Compare worst-case insert between Vec and LinkedList
While the documented worst-case insert of a LinkedList is O(n/2), which
seems better than Vec’s amortized worst-case cost of O(n), benchmarking
outlines that the LinkedList is still many orders of magnitude slower
than Vec in a worst-case insertion.

The likely explanation is that a LinkedList’s pointer indirection is
slower than a Vec’s sequential access.