Initialize tree array structure
The goal of a tree array is more efficiency, average O(log n), than both
LinkedList (O(min(i, n-i))) and Vec (O(n-i)) at insert operations.

The implementation we will pick is that of a splay tree which holds the
number of elements of each node's subtree in the node, allowing for fast
computation of a node's index.

Get operations will go from Vec's O(1) to O(log n), but we estimate that
this won't be a problem in practice, in the particular use-case of
multiuser text editing.